producer/requirements.txt
kafka-python==2.0.2
faker==13.12.0
loguru==0.6.0
